### 2018-10-10 #1009 @arfon
@zkuralt - please take at look at the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) in our submission requirements and format the YAML header of your paper the same way. Also, please note that YAML is white-space sensitive when making these changes.

Once you've made your changes you can ask @whedon to recompile your paper with the command `@whedon generate pdf`
