### 2018-10-15 #1009 @amoeba
Hi @trallard 	I could probably commit to getting a review on by November 15th if that works for you. Please let me know and I can plan on it.
