### 2018-11-04 #1064 @whedon
[REVIEW]: wolfexplorer: a tool for visualization and exploration of complex multi-year multi-specimen datasets
**Submitting author:** @zkuralt (<a href="http://orcid.org/0000-0001-7201-4282">Žan Kuralt</a>)
**Repository:** <a href="https://github.com/romunov/wolfexplorer" target ="_blank">https://github.com/romunov/wolfexplorer</a>
**Version:** v0.9
