### 2018-11-13 romunov/wolfexplorer#42 @amoeba
JOSS review
Hi @romunov 	and @zkuralt , JOSS encourages reviewers to note and discuss review issues on the issue tracker for the software being reviewed so I've created this issue here.

**Review: Accept with minor revisions**

This is a really interesting Shiny application and I applaud your attention to detail. I found the installation instructions in the README easy to follow and accurate. The Shiny application itself looks clean, works well, and feels intuitive. Overall, this has been one of my easiest reviews so thank you for taking the time to get things on order.

I found a few issues I'd like you to address before I change my review to an Accept. Please take a look and let me know if you agree or disagree on any of them. I've also put some notes that I don't need you to address for JOSS but think might improve your software a bit.

#### Review items

- **Version: Does the release version given match the GitHub release (v0.9)?**

  No, there are no releases for this software. Please make a release on GitHub.

- **Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?**

  I didn't see any automated tests, but I think your instructions in the REAMDE are sufficient to check this item off (so I did). You might consider adding some unit tests for the R code or even some headless tests for the Shiny app at some later point.

- **Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**

  I didn't see these in the README. Please add these sections.

- **Paper**

  The reference for split-apply-combine paper needs its DOI of [https://doi.org/10.18637/jss.v040.i01](https://doi.org/10.18637/jss.v040.i01). Please add it and re-generated the paper proof.

#### Non-review notes:

_You don't need to address these for me to sign off on the review_

- In README, I see the text "Luckily WGS (EPSG: 4326)". I think this should read, "Luckily WGS 84 (EPSG: 4326) to be more accurate.
- While the installation instructions listed the packages I need to get the application running, I wonder if you could also put the dependencies in the `DESCRIPTION` file to make the dependencies more explicit. This would (1) make it possible to install all of the dependencies with `devtools::install_deps()` and (2) would allow external tools to track the reverse dependency.
