### 2018-11-13 #1064 @amoeba
Hi @trallard I've made my review (Sorry its one day late). I accept with minor revisions. I left some notes for the authors on https://github.com/romunov/wolfexplorer/issues/42 but checked most everything off. One thing I didn't know about was whether the paper should go in a 	`joss` subfolder or the usual `paper` subfolder at the root of the repo. Do you know?

For now, I'll wait to hear back from the authors on my review and you on the above `paper.md` location question. Thanks!
