### 2018-11-13 romunov/wolfexplorer#42 @romunov
Thank you @amoeba for your thorough review. Not only do you raise an issue but also offer a suggestion on how to solve it. In my opinion, this is very important in a review process and many could learn from you.

I think we can address all of the valid points you raised. You will hear from us sooner than later.
