### 2018-11-16 a80f44e
Work on reviewer comments for JOSS (#42)
