+++ joss/paper.bib
@@ -117,6 +117,7 @@ Plots},
     title = {The Split-Apply-Combine Strategy for Data Analysis},
     author = {Hadley Wickham},
     journal = {Journal of Statistical Software},
+    doi = {10.18637/jss.v040.i01},
     year = {2011},
     volume = {40},
     number = {1},
