### 2018-11-16 #1009 @zkuralt
I suspect that whedon generates PDF from master branch. We will generate it again after merging joss_review branch to master.
