### 2018-11-16 romunov/wolfexplorer#42 @zkuralt
I created a new branch "joss_review" for revised version. @amoeba we will let you know once we check all the review comments.

Cheers!
