### 2018-11-16 #1009 @trallard
Yes by default it uses master

I can see the paper correctly unless you have been working on an alternate branch for corrections and stuff @zkuralt
