### 2018-11-16 #1064 @trallard
hey @amoeba  thanks for the speedy review

I have looked at . a number of accepted submissions and it seems that there is a mixture of approaches as where to place the paper, some packages have it under a `paper` directory and other under `joss` I think that since the final paper is archived within JOSS itself this _is not a blocker_

---
@zkuralt  I am  going to recompile the paper here as the pre-review issue is closed now
