### 2018-11-16 #1064 @zkuralt
@arfon Thanks, I missed it when adding DOI to the bibliography item.
