### 2018-11-25 romunov/wolfexplorer#42 @romunov
@amoeba I have looked into automatic testing of shiny apps. A package called `shinytest` does exactly that, however due to some bugs in its code, it does not work properly with applications implemented through `shinydashboard` package yet. This means that tests would have to be coded manually and some restructuring of the code would be necessary.

Right now, the README/"vignette" describes the steps of importing accompanying data and offers screenshots (basically what shinytest does). Would this satisfy your criterion you describe in point 2 of your review?
