### 2018-11-28 romunov/wolfexplorer#42 @amoeba
> Would this satisfy your criterion you describe in point 2 of your review?

Yes. @romunov 	that'd be fine.
