### 2018-12-01 romunov/wolfexplorer#42 @romunov
@amoeba I believe we've addressed all the issues you've outlined in your review. Let us know what you think.
