### 2018-12-03 romunov/wolfexplorer#42 @amoeba
Hi @romunov, thanks for taking the time to address the issues I've outlined. I took another look and everything looks in order. I'm going to close this and leave a comment for @trallard on https://github.com/openjournals/joss-reviews/issues/1064. Thanks!
