### 2018-12-03 #1064 @amoeba
Hi @trallard, all review checklist items have been satisfactorily met by the authors and I now change my review to Accept.
