### 2018-12-03 #1064 @trallard
@zkuralt since @amoeba has now recommended the acceptance of the package we need you to complete the following actions:
- [  ] confirm that the version of the paper in the master branch is the final version
- [  ] make a new release of the package, upload the revised software to your DOI-granting data/software repository, and post the DOI here
