### 2018-12-03 #1064 @romunov
I have confirmed that the latest release corresponds to what has been put up for review and includes subsequent changes suggested by the reviewer.

Source code of v0.9 release has been deposited in Zenodo under DOI [10.5281/zenodo.1890359](https://zenodo.org/record/1890359).
