### 2018-12-10 #1064 @trallard
@arfon: this submission is accepted and ready to be published 🎉👾

@amoeba thank you very much for your time and valuable contribution to JOSS as a reviewer for this submission 🙌🏻
