### 2018-12-10 ba01948
Add DOI to README
