+++ R/file_input.R
@@ -1,62 +1,36 @@
 inputFileSamples <- reactive({
-  x <- input$data_samples
-  if (is.null(x)) {
-    data.frame(lng = NA, lat = NA, date = NA, sample_type = NA, animal = NA, sex = NA, 
-               sample_name = NA, id = NA, reference_sample = NA)[0, ]
-  } else {
-    x <- tryCatch(fread(x$datapath, 
-                        encoding = "UTF-8",
-                        colClasses = c("numeric", "numeric", "character", "character", "character", 
-                                       "character", "character", "character"),
-                        data.table = FALSE),
-                  error = function(e) e,
-                  warning = function(w) w
-    )
-    
-    if (any(class(x) %in% c("simpleWarning", "simpleError"))) {
-      alert("Input data not formatted properly. Please compare your input file to the specs.")
-      x <- data.frame(lng = NA, lat = NA, date = NA, sample_type = NA, animal = NA, sex = NA, 
-                      sample_name = NA, id = NA, reference_sample = NA)[0, ]
-      return(x)
-    }
-    
-    validate(
-      need(all(colnames(x) %in% c("x", "y", "date", "sample_type", "animal", "sex", "sample_name", 
-                                  "reference_sample")), 
-           "Column names not as expected.")
-    )
-    
-    x <- GKtoWGS(x)
-    x$date <- as.Date(x$date, format = "%Y-%m-%d")
-    x$id <- 1:nrow(x)
-    
-    x
+  xy <- dbReadTable(conn = db, "samples")
+  # x         y       date sample_type animal sex sample_name reference_sample
+  # 1   415290.5  48492.58 2010-04-23      Saliva    657   M     AH.03MT          AH.03MT
+  # 2   439351.3  44444.31 2014-12-22      Saliva    657   M     EX.1JKT          AH.03MT
+  # 3   445348.9  44420.60 2014-12-15      Saliva    658   M     EX.1JJ1          AL.05PH
+  
+  validate(
+    need(all(colnames(xy) %in% c("x", "y", "date", "sample_type", "animal", "sex", 
+                                 "sample_name", "reference_sample")), 
+         "Column names not as expected.")
+  )
+  
+  if (nrow(xy) > 0) {
+    xy <- GKtoWGS(xy)
+    xy$date <- as.Date(xy$date, format = "%Y-%m-%d")  # sqlite can't handle dates properly
+    xy$id <- 1:nrow(xy)
   }
+  
+  xy
 })
 
 inputFileParentage <- reactive({
-  x <- input$data_parentage
-  if (is.null(x)) {
-    data.frame(offspring = NA, mother = NA, father = NA, cluster = NA)[0, ]
-  } else {
-    out <- tryCatch(fread(x$datapath, encoding = "UTF-8", 
-                          colClasses = c("character", "character", "character", "character"),
-                          data.table = FALSE),
-                    error = function(e) e,
-                    warning = function(w) w
-    )
-    
-    if (any(class(out) %in% c("simpleWarning", "simpleError"))) {
-      alert("Input data not formatted properly. Please compare your input file to the specs.")
-      out <- data.frame(offspring = NA, mother = NA, father = NA, cluster = NA)[0, ]
-      return(out)
-    } 
-    
-    # if column names do not match predefined form, warn user
-    validate(
-      need(all(colnames(out) %in% c("offspring", "mother", "father", "cluster")),
-           "Column names not as expected.")
-    )
-    return(out)
-  }
+  xy <- dbReadTable(conn = db, "parentage")
+  # offspring  mother  father cluster
+  # 1     M2122 AU.0AEF AH.03MT       2
[…]
