### 2019-04-06 956db0c
Pretty code, remov print statements, translat some comments to Slovene
